Restructure of fetch
Restructure of fetch util files for convenience
