+++ src/auth/src/main/resources/static/js/utils/fetch/fetch-util.js
@@ -1,0 +1,3 @@
+/**
+ * 
+ */
