+++ src/auth/src/main/resources/static/js/utils/fetch/implementations/fetch-class.js
@@ -1,0 +1,16 @@
+"use strict";
+
+class FetchUtil {
+	constructor(postImpl, getHtmlImpl) {
+		this.postImpl = postImpl;
+		this.getHtmlImpl = getHtmlImpl;
+	}
+	
+	post() {
+		this.postImpl();
+	}
+	
+	getHtml() {
+		this.getHtmlImpl();
+	}
+}
