+++ src/auth/src/main/resources/static/js/utils/fetch/implementations/generic-implementations/fetch-generic.js
