+++ src/auth/src/main/resources/static/js/utils/fetch/implementations/generic-implementations/methods/get-html.js
