+++ src/auth/src/main/resources/static/js/utils/fetch/implementations/jwt-implementations/fetch-jwt.js
